Moved `AnalysisState`.
