+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/AnalysisController.java
@@ -348,35 +348,6 @@ public class AnalysisController {
 	}
 
 	/**
-	 * The current state of the analysis.
-	 *
-	 * @author Christoph Michelbach
-	 */
-	public enum AnalysisState {
-		/**
-		 * The analysis is currently running.
-		 */
-		RUNNING,
-
-		/**
-		 * Stopping the analysis without trying to preserve data accumulated in the
-		 * current phase.
-		 */
-		ABORTING,
-
-		/**
-		 * Stopping the analysis but trying to preserve data accumulated in the current
-		 * phase.
-		 */
-		ENDING,
-
-		/**
-		 * The analysis isn't running.
-		 */
-		TERMINATED
-	}
-
-	/**
 	 * The callback the {@link AnalysisController} uses to interrupt the working thread if
 	 * the user is pausing or aborting the analysis.
 	 *
