+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/facade/BeagleController.java
@@ -1,6 +1,7 @@
 package de.uka.ipd.sdq.beagle.core.facade;
 
 import de.uka.ipd.sdq.beagle.core.AnalysisController;
+import de.uka.ipd.sdq.beagle.core.AnalysisState;
 import de.uka.ipd.sdq.beagle.core.Blackboard;
 import de.uka.ipd.sdq.beagle.core.LaunchConfiguration;
 import de.uka.ipd.sdq.beagle.core.ProjectInformation;
@@ -129,7 +130,7 @@ public class BeagleController {
 	 *
 	 */
 	public void pauseAnalysis() {
-		this.analysisController.setAnalysisState(AnalysisController.AnalysisState.ENDING);
+		this.analysisController.setAnalysisState(AnalysisState.ENDING);
 	}
 
 	/**
@@ -138,7 +139,7 @@ public class BeagleController {
 	 *
 	 */
 	public void continueAnalysis() {
-		this.analysisController.setAnalysisState(AnalysisController.AnalysisState.RUNNING);
+		this.analysisController.setAnalysisState(AnalysisState.RUNNING);
 	}
 
 	/**
@@ -148,7 +149,7 @@ public class BeagleController {
 	 *
 	 */
 	public void abortAnalysis() {
-		this.analysisController.setAnalysisState(AnalysisController.AnalysisState.ABORTING);
+		this.analysisController.setAnalysisState(AnalysisState.ABORTING);
 	}
 
 	/**
